+++ CHANGELOG.md
@@ -27,8 +27,9 @@ According to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) , the `Unr
 - [#465](https://github.com/FC4E-CAT/fc4e-cat-api/pull/465) CAT-822 Assessment type view and fill put order in criteria.
 - [#467](https://github.com/FC4E-CAT/fc4e-cat-api/pull/467) CAT-832 Add 127.0.0.1 to Keycloak Public Client to Accept Requests.
 - [#468](https://github.com/FC4E-CAT/fc4e-cat-api/pull/468) CAT-827 CAT-829 CAT-828 Publish Assessment in Zenodo / Store zenodo info in database/ Notifications to creator.
-- [#469](https://github.com/FC4E-CAT/fc4e-cat-api/pull/469) CAT-830 Publish deposit to Zenodo
+- [#469](https://github.com/FC4E-CAT/fc4e-cat-api/pull/469) CAT-834 Publish deposit to Zenodo
 - [#470](https://github.com/FC4E-CAT/fc4e-cat-api/pull/470) CAT-836 Get deposit info from Zenodo 
+- [#471](https://github.com/FC4E-CAT/fc4e-cat-api/pull/471) CAT-830 Configure zenodo feature in cat
 
 
 ## 1.9.2 - 2025-02-26